changed: split out non-typetag dependent code from WellInterface
